Add URL handling and file size retrieval functions
