+++ lib/utils/string_extension.dart
